add osmo_use_count API
Provide a common implementation of use counting that supports naming each user
as well as counting more than just one use per user.

In osmo-msc, we were originally using a simple int counter to see whether a
connection is still in use or should be discarded. For clarity, we later added
names to each user in the form of a bitmask of flags, to figure out exactly
which users are still active: for logging and to debug double get / double put
bugs. This however is still not adequate, since there may be more than one CM
Service Request pending. Also, it is a specialized implementation that is not
re-usable. In the course of the inter-MSC Handover related refactoring, I would
like to fix the problem of inadequate counting of multiple concurrent CM
Service Requests, and would also like to add more than one use count in
osmo-msc. A standardized implementation is desirable.

Change-Id: Ife31e6798b4e728a23913179e346552a7dd338c0